Create a new session on 404 when refreshing (closes #5)
The consul agent can report a false negative serfHealth status.
When that happens the consul server removes the faulty member.
This has two consequences :
1 - all the sessions associated with the node are deleted
2 - the creation of a new session might still fail with
500,Internal Server Error, Check 'serfHealth' is in critical state

This commit is an attempt to be more resilient in this cases and keep
trying to create a new session. If it succeded the system will
keep working as if nothing happened, otherwise after the max num of retry
the constructr machine will terminate the akka system.

The default constructr machines configuration don't give consul enough
time to recover under these circumstances. I suggest to change
the configuration for production env like to something like :

  coordination-timeout = 10 seconds
  nr-of-retries        = 10
  refresh-interval     = 60 seconds
  retry-delay          = 10 seconds
  ttl-factor           = 5.0